change nav and create login page and register page
frame work update
