+++ roster_management_system/lib/widge/login.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+class loginpage extends StatelessWidget {
+  const loginpage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
