+++ roster_management_system/lib/widge/register.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+class registerpage extends StatelessWidget {
+  const registerpage({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
