+++ roster_management_system/pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   firebase: ^9.0.2
+  router_management: ^1.3.3
 
 dev_dependencies:
   flutter_test:
